refactor: rename computation_engine_lib to engine (Phase 10, #625)
