+++ custom_components/localshift/engine/excess_solar_signals.py
