+++ custom_components/localshift/engine/pattern_analyzer.py
