+++ custom_components/localshift/engine/slot_builder.py
