+++ custom_components/localshift/engine/soc_simulator.py
