+++ tests/test_optimizer_dp_solve.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 
 import pytest
 
-from custom_components.localshift.computation_engine_lib.optimizer_dp import (
+from custom_components.localshift.engine.optimizer_dp import (
     DPPlanner,
     ObjectiveTerms,
     OptimizerConfig,
