+++ tests/test_optimizer_facade.py
@@ -2,7 +2,7 @@
 
 from datetime import UTC, datetime
 
-from custom_components.localshift.computation_engine_lib.optimizer_facade import (
+from custom_components.localshift.engine.optimizer_facade import (
     OptimizerFacade,
 )
 from custom_components.localshift.coordinator import CoordinatorData
