+++ tests/test_optimizer_runner.py
@@ -1,3 +1,3 @@
 # Wrapper for TDD compliance - test file location tracked in GitHub issue
 # TODO: Consolidate test file locations (either move or update TDD hook)
-from tests.computation_engine_lib.test_optimizer_runner import *  # noqa: F401, F403
+from tests.engine.test_optimizer_runner import *  # noqa: F401, F403
