+++ tests/test_optimizer_self_consumption.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from custom_components.localshift.computation_engine_lib.optimizer_dp import (
+from custom_components.localshift.engine.optimizer_dp import (
     DPPlanner,
     OptimizerConfig,
     OptimizerInputs,
