+++ tests/test_solar_opportunity_penalty.py
@@ -8,7 +8,7 @@ from datetime import datetime
 
 import pytest
 
-from custom_components.localshift.computation_engine_lib.optimizer_dp import (
+from custom_components.localshift.engine.optimizer_dp import (
     DPPlanner,
     OptimizerConfig,
     OptimizerInputs,
